Preliminary support for shipment notifications (Fedex only).
Create a new object to represent the notifications specification. In
theory this object is could be used to configure any service. Obviously
it's design is heavily based on the needs of Fedex and will require
further refinement if other services are supported.

But despite it just being a start it does provide some abstraction
from the underlying API. We are currently on v13 but if ActiveShipping
moves to the latest version for example the XML changes in insignificant
but incompatible ways. The notification object will abstract us from
those changes.

Currently the notification object defines the possible values and
provides some validation. We may need to move that out of the notification
object and into the carrier as the notification object becomes more generic
but I figure error early for now as we may never get support of other
carriers.